+++ api/rutas-api/Pipfile
@@ -7,6 +7,7 @@ name = "routes-api"
 flask = "*"
 flask-sqlalchemy = "*"
 psycopg2-binary = "*"
+requests = "*"
 wheel = "*"
 marshmallow = "*"
 python-abc = "*"
